unicode: Split out pure unicode bits from //idna
These are generated from different specifications and are required in
different places, even if idna depends on parts of unicode.
